chore(compass-components, compass-databases-navigation): Move sidebar navigation to its own package
