+++ packages/compass-components/package.json
@@ -40,7 +40,6 @@
     "@leafygreen-ui/card": "^5.1.3",
     "@leafygreen-ui/checkbox": "^6.0.5",
     "@leafygreen-ui/confirmation-modal": "^2.2.1",
-    "@leafygreen-ui/emotion": "^4.0.0",
     "@leafygreen-ui/icon": "^11.4.0",
     "@leafygreen-ui/icon-button": "^9.1.6",
     "@leafygreen-ui/leafygreen-provider": "^2.1.2",
@@ -60,7 +59,6 @@
     "@leafygreen-ui/typography": "^8.0.2",
     "@mongodb-js/compass-logging": "^0.4.0",
     "@react-aria/interactions": "^3.6.0",
-    "@react-aria/visually-hidden": "^3.2.3",
     "@react-aria/utils": "^3.9.0"
   },
   "peerDependencies": {
