+++ packages/compass-components/src/index.ts
@@ -60,4 +60,10 @@ export {
   ResizeHandle,
   ResizeDirection,
 };
-export { DatabasesNavigationTree } from './components/databases-navigation-tree';
+export {
+  useFocusState,
+  useHoverState,
+  FocusState,
+} from './hooks/use-focus-hover';
+export { useTree, useTreeItem } from './hooks/use-navigation-tree';
+export { ContentWithFallback } from './components/content-with-fallback';
