+++ packages/compass-databases-navigation/.eslintignore
@@ -1,0 +1,2 @@
+.nyc-output
+dist
