+++ packages/compass-databases-navigation/src/index.ts
@@ -1,0 +1,1 @@
+export { DatabasesNavigationTree as default } from './databases-navigation-tree';
