+++ packages/compass-sidebar/package.json
@@ -52,6 +52,7 @@
     "@mongodb-js/compass-app-stores": "^5.15.0",
     "@mongodb-js/compass-components": "^0.7.0",
     "@mongodb-js/compass-connect": "^7.15.0",
+    "@mongodb-js/compass-databases-navigation": "^0.1.0",
     "@mongodb-js/compass-deployment-awareness": "^11.15.0",
     "@mongodb-js/compass-server-version": "^5.15.0",
     "@mongodb-js/compass-ssh-tunnel-status": "^6.15.0",
@@ -74,7 +75,6 @@
     "font-awesome": "^4.7.0",
     "hadron-app": "^4.15.0",
     "hadron-app-registry": "^8.5.0",
-    "hadron-ipc": "^2.6.0",
     "hadron-react-bson": "^5.6.0",
     "hadron-react-buttons": "^5.5.0",
     "html-webpack-plugin": "^3.2.0",
@@ -115,7 +115,6 @@
     "react-ios-switch": "^0.1.19",
     "react-redux": "^5.0.6",
     "react-tooltip": "^3.11.1",
-    "react-virtualized": "^9.21.0",
     "redux": "^3.7.2",
     "redux-thunk": "^2.3.0",
     "reflux": "^0.4.1",
