+++ packages/databases-collections/package.json
@@ -16,12 +16,12 @@
     "start": "webpack-dev-server --config ./config/webpack.dev.config.js",
     "start:watch": "npm run clean && webpack --config ./config/webpack.watch.config.js",
     "prestart": "electron-rebuild --force --only keytar",
+    "pretest": "npm run compile",
     "test": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\"",
     "test:watch": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\" --watch",
     "test:karma": "xvfb-maybe cross-env NODE_ENV=test karma start",
     "cover": "nyc npm run test",
     "test-check-ci": "npm run cover && npm run test:karma && npm run check",
-    "precheck": "npm run compile",
     "check": "npm run lint && npm run depcheck",
     "prepublishOnly": "npm run compile",
     "lint": "eslint \"./src/**/*.{js,jsx}\" \"./test/**/*.js\" \"./electron/**/*.js\" \"./config/**/*.{js,jsx}\"",
